Code Refactored for Hotel service
