+++ hotel-service/src/main/java/com/epam/hotel/HotelApplication.java
@@ -1,4 +1,4 @@
-package com.epam.hotelservice;
+package com.epam.hotel;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 
 @SpringBootApplication
 @EnableEurekaClient
-public class HotelServiceApplication {
+public class HotelApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(HotelServiceApplication.class, args);
+		SpringApplication.run(HotelApplication.class, args);
 	}
 
 }
