+++ hotel-service/src/main/java/com/epam/hotel/controller/HotelController.java
@@ -1,13 +1,13 @@
-package com.epam.hotelservice.controller;
+package com.epam.hotel.controller;
 
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import com.epam.hotelservice.entity.Hotel;
+import com.epam.hotel.entity.Hotel;
 
 @RequestMapping("/v1/api")
-public interface HotelServiceControllerInterface {
+public interface HotelController {
 	@PostMapping("/hotels")
 	public Hotel addHotel(@RequestBody Hotel hotel);
 }
