+++ hotel-service/src/main/java/com/epam/hotel/entity/Hotel.java
@@ -1,10 +1,15 @@
-package com.epam.hotelservice.entity;
+package com.epam.hotel.entity;
+
+import java.util.List;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToMany;
+import javax.persistence.OneToOne;
 import javax.persistence.Table;
 
 import lombok.AllArgsConstructor;
@@ -12,20 +17,21 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "ROOM_TYPE")
+@Table(name = "HOTEL")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class RoomType {
+public class Hotel {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private int id;
-	@Column(name = "ROOM_TYPE")
-	private String roomName;
-	@Column(name = "RENT_PER_DAY")
-	private float rentPerDay;
-	@Column(name = "ROOM_CAPACITY")
-	private int roomCapacity;
+	@Column(name = "HOTEL_NAME")
+	private String hotelName;
+	@OneToOne
+	@JoinColumn(name = "HOTEL_ADDRESS_ID")
+	private HotelAddress hotelAddress;
 	@Column(name = "IS_ACTIVE")
-	private boolean isActive;
+	private Boolean isActive;
+	@OneToMany(mappedBy = "hotel")
+	private List<Room> room;
 }
