+++ hotel-service/src/main/java/com/epam/hotel/service/HotelClass.java
@@ -1,0 +1,5 @@
+package com.epam.hotel.service;
+
+public class HotelClass {
+
+}
