--- hotel-service/src/main/java/com/epam/hotelservice/service/HotelServiceClass.java
@@ -1,5 +1,0 @@
-package com.epam.hotelservice.service;
-
-public class HotelServiceClass {
-
-}
